create groups to separate files by type separate item image to reuse create DetailView add navigation to DetailView with sharing of the selected Item instance
